Addressing feedback from Siva
